Add a systemd unit file for openport
Currently start/stop is pretty inconsistent with openport and status
is meaningless.

By moving to a unit file and letting systemd be the parent of a
non-daemonized openport, we get start/stop/status, but also, journal
works as expected, which is quite nice.

This unit file auto-restarts which is useful.

Signed-off-by: Phil Dibowitz <phil@ipom.com>